jd2018: config.fs: init: Add AID_VENDOR_QRTR for qrtr-ns and remove unneeded flag
Add the AID_VENDOR_QRTR AID in config.fs. Change qrtr-ns to use this AID
for system and group. This AID is reserved for applications that
perform control operations on the QRTR socket family.

The QRTR node ID is hardcoded to 1 within the kernel. Do not configure
the node from userspace. This allows qrtr-ns to drop NET_ADMIN
capabilities.

Change-Id: If2f20bdb8e591c74bed5da29e0c2d3aefb6dc926
Reviewed-on: https://mokeedev.review/c/MoKee/android_device_lenovo_jd2018/+/47120
Tested-by: Art_Chen <1263972314@qq.com>
Reviewed-by: Art_Chen <1263972314@qq.com>